+++ scandeval/benchmarks/dkhate.py
@@ -4,7 +4,7 @@ from datasets import Dataset
 from typing import Tuple
 import logging
 
-from .abstract import BinaryClassificationBenchmark
+from .abstract import TextClassificationBenchmark
 from ..utils import doc_inherit
 from ..datasets import load_dataset
 
@@ -12,7 +12,7 @@ from ..datasets import load_dataset
 logger = logging.getLogger(__name__)
 
 
-class DkHateBenchmark(BinaryClassificationBenchmark):
+class DkHateBenchmark(TextClassificationBenchmark):
     '''Benchmark of language models on the DKHate dataset.
 
     Args:
@@ -44,8 +44,14 @@ class DkHateBenchmark(BinaryClassificationBenchmark):
                  cache_dir: str = '.benchmark_models',
                  evaluate_train: bool = False,
                  verbose: bool = False):
+        id2label = ['NOT', 'OFF']
+        label_synonyms = [
+            ['LABEL_0', id2label[0]],
+            ['LABEL_1', id2label[1]],
+        ]
         super().__init__(name='dkhate',
-                         id2label=['NOT', 'OFF'],
+                         id2label=id2label,
+                         label_synonyms=label_synonyms,
                          cache_dir=cache_dir,
                          evaluate_train=evaluate_train,
                          verbose=verbose)