+++ scandeval/benchmarks/twitter_sent.py
@@ -4,7 +4,7 @@ from datasets import Dataset
 from typing import Tuple
 import logging
 
-from .abstract import SentimentClassificationBenchmark
+from .abstract import TextClassificationBenchmark
 from ..utils import doc_inherit
 from ..datasets import load_dataset
 
@@ -12,7 +12,7 @@ from ..datasets import load_dataset
 logger = logging.getLogger(__name__)
 
 
-class TwitterSentBenchmark(SentimentClassificationBenchmark):
+class TwitterSentBenchmark(TextClassificationBenchmark):
     '''Benchmark of language models on the TwitterSent dataset.
 
     Args:
@@ -44,7 +44,15 @@ class TwitterSentBenchmark(SentimentClassificationBenchmark):
                  cache_dir: str = '.benchmark_models',
                  evaluate_train: bool = False,
                  verbose: bool = False):
+        id2label = ['negative', 'neutral', 'positive']
+        label_synonyms = [
+            ['LABEL_0', 'negativ', 'neikvætt', 'Negative', id2label[0]],
+            ['LABEL_1', 'nøytral', 'hlutlaus', 'Neutral', id2label[1]],
+            ['LABEL_2', 'positiv', 'jákvætt', 'Positive', id2label[2]]
+        ]
         super().__init__(name='twitter-sent',
+                         id2label=id2label,
+                         label_synonyms=label_synonyms,
                          cache_dir=cache_dir,
                          evaluate_train=evaluate_train,
                          verbose=verbose)